backend/requirements.txt
tqdm
numpy==1.26.4
pandas==2.2.2
scikit-learn==1.4.2
nltk==3.9.1
sentence-transformers==3.0.1
faiss-cpu==1.8.0
PyPDF2==3.0.1
python-pptx==1.0.2
openai>=1.0.0
python-dotenv==1.0.1
scipy==1.11.4
joblib==1.3.2
rank_bm25==0.2.2
psycopg[binary]==3.1.19
pgvector==0.3.4
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-jose[cryptography]==3.3.0
bcrypt>=3.2.0
python-multipart==0.0.9
